feat: implemented all custom exceptions and their handlers

+++ booking/src/main/java/com/komsije/booking/service/AccommodationServiceImpl.java
@@ -2,6 +2,7 @@ package com.komsije.booking.service;
 
 import com.komsije.booking.dto.AccommodationDto;
 import com.komsije.booking.dto.AvailabilityDto;
+import com.komsije.booking.exceptions.ElementNotFoundException;
 import com.komsije.booking.mapper.AccommodationMapper;
 import com.komsije.booking.model.Accommodation;
 import com.komsije.booking.model.AccommodationType;
@@ -12,11 +13,12 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 public class AccommodationServiceImpl implements AccommodationService {
     @Autowired
-    private  AccommodationMapper mapper;
+    private AccommodationMapper mapper;
     private final AccommodationRepository accommodationRepository;
 
     @Autowired
@@ -24,19 +26,13 @@ public class AccommodationServiceImpl implements AccommodationService {
         this.accommodationRepository = accommodationRepository;
     }
 
-    public AccommodationDto findById(Long id) {
-        try {
-            return mapper.toDto(accommodationRepository.findById(id).orElseGet(null));
-        }
-        catch (NullPointerException e){
-            return null;
-        }
+    public AccommodationDto findById(Long id) throws ElementNotFoundException {
+        return mapper.toDto(accommodationRepository.findById(id).orElseThrow(()->new ElementNotFoundException("Element with given ID doesn't exist!")));
     }
 
 
-
     public List<AccommodationDto> findAll() {
-       return mapper.toDto(accommodationRepository.findAll());
+        return mapper.toDto(accommodationRepository.findAll());
     }
 
     public AccommodationDto save(AccommodationDto accommodationDto) {
@@ -45,19 +41,20 @@ public class AccommodationServiceImpl implements AccommodationService {
     }
 
     @Override
-    public AccommodationDto update(AccommodationDto accommodationDto) {
-        Accommodation accommodation = accommodationRepository.findById(accommodationDto.getId()).orElseGet(null);
-        if (accommodation == null){
-            return null;
-        }
-//        Accommodation updatedAccommodation = mapper.fromDto(accommodationDto);
-        mapper.update(accommodation,accommodationDto);
+    public AccommodationDto update(AccommodationDto accommodationDto) throws ElementNotFoundException {
+        Accommodation accommodation = accommodationRepository.findById(accommodationDto.getId()).orElseThrow(()->new ElementNotFoundException("Element with given ID doesn't exist!"));
+        mapper.update(accommodation, accommodationDto);
         accommodationRepository.save(accommodation);
         return accommodationDto;
     }
 
-    public void delete(Long id) {
-        accommodationRepository.deleteById(id);
+    public void delete(Long id) throws ElementNotFoundException {
+        if (accommodationRepository.existsById(id)){
+            accommodationRepository.deleteById(id);
+        }else{
+            throw new ElementNotFoundException("Element with given ID doesn't exist!");
+        }
+
     }
 
     public List<AccommodationDto> getByAccommodationType(AccommodationType type) {
@@ -65,35 +62,24 @@ public class AccommodationServiceImpl implements AccommodationService {
     }
 
     @Override
-    public AccommodationDto updateAvailability(Long accommodationId, AvailabilityDto availabilityDto) {
-        Accommodation accommodation = accommodationRepository.findById(accommodationId).orElseGet(null);
-        if (accommodation == null){
-            return null;
-        }
+    public AccommodationDto updateAvailability(Long accommodationId, AvailabilityDto availabilityDto) throws ElementNotFoundException {
+        Accommodation accommodation = accommodationRepository.findById(accommodationId).orElseThrow(()->new ElementNotFoundException("Element with given ID doesn't exist!"));
         mapper.update(accommodation, availabilityDto);
         accommodationRepository.save(accommodation);
         return mapper.toDto(accommodation);
     }
 
-    public List<AccommodationDto> getByLocationNumOfGuestsAndDate(String location, Integer numOfGuests, LocalDateTime startDate, LocalDateTime endDate){
-        return mapper.toDto(accommodationRepository.getAccommodationsByLocationNumOfGuestsAndDate(location,numOfGuests, startDate, endDate));
+    public List<AccommodationDto> getByLocationNumOfGuestsAndDate(String location, Integer numOfGuests, LocalDateTime startDate, LocalDateTime endDate) {
+        return mapper.toDto(accommodationRepository.getAccommodationsByLocationNumOfGuestsAndDate(location, numOfGuests, startDate, endDate));
     }
 
-    public List<AccommodationDto> getByAmenities(List<String> amenities){
+    public List<AccommodationDto> getByAmenities(List<String> amenities) {
         return mapper.toDto(accommodationRepository.getAccommodationsByAmenities(amenities));
     }
 
-    public Accommodation findModelById(Long id){
-        try {
-            return accommodationRepository.findById(id).orElseGet(null);
-        }
-        catch (Exception e){
-            return null;
-        }
+    public Accommodation findModelById(Long id) throws ElementNotFoundException {
+        return accommodationRepository.findById(id).orElseThrow(()->new ElementNotFoundException("Element with given ID doesn't exist!"));
     }
-
-
-
 
 
 }
