+++ booking/src/main/java/com/komsije/booking/service/ReviewServiceImpl.java
@@ -1,6 +1,7 @@
 package com.komsije.booking.service;
 
 import com.komsije.booking.dto.ReviewDto;
+import com.komsije.booking.exceptions.ElementNotFoundException;
 import com.komsije.booking.mapper.ReviewMapper;
 import com.komsije.booking.model.Review;
 import com.komsije.booking.repository.ReviewRepository;
@@ -21,13 +22,8 @@ public class ReviewServiceImpl implements ReviewService {
         this.reviewRepository = reviewRepository;
     }
 
-    public ReviewDto findById(Long id) {
-        try {
-            return mapper.toDto(reviewRepository.findById(id).orElseGet(null));
-        }
-        catch (NullPointerException e){
-            return null;
-        }
+    public ReviewDto findById(Long id) throws ElementNotFoundException {
+        return mapper.toDto(reviewRepository.findById(id).orElseThrow(() ->  new ElementNotFoundException("Element with given ID doesn't exist!")));
     }
 
     public List<ReviewDto> findAll() {
@@ -40,29 +36,28 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
-    public ReviewDto update(ReviewDto reviewDto) {
-        Review review = reviewRepository.findById(reviewDto.getId()).orElseGet(null);
-        if (review == null) {
-            return null;
-        }
+    public ReviewDto update(ReviewDto reviewDto) throws ElementNotFoundException {
+        Review review = reviewRepository.findById(reviewDto.getId()).orElseThrow(() ->  new ElementNotFoundException("Element with given ID doesn't exist!"));
         mapper.update(review, reviewDto);
         reviewRepository.save(review);
         return reviewDto;
     }
 
-    public void delete(Long id) {
-        reviewRepository.deleteById(id);
+    public void delete(Long id) throws ElementNotFoundException {
+        if (reviewRepository.existsById(id)){
+            reviewRepository.deleteById(id);
+        }else{
+            throw new ElementNotFoundException("Element with given ID doesn't exist!");
+        }
+
     }
 
     public List<ReviewDto> getApprovedReviews() {
         return mapper.toDto(reviewRepository.getReviewsByIsApprovedIsTrue());
     }
 
-    public void setApproved(Long id){
-        Review review = reviewRepository.findById(id).orElseGet(null);
-        if(review==null){
-            return;
-        }
+    public void setApproved(Long id) throws ElementNotFoundException {
+        Review review = reviewRepository.findById(id).orElseThrow(() ->  new ElementNotFoundException("Element with given ID doesn't exist!"));
         review.setApproved(true);
         reviewRepository.save(review);
     }
